requirements.txt
psycopg2
#psycopg2.extras
progressbar2
